fix crud aritcle et refonte des route api

+++ routes/api.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\API\InterestPointsController;
 use App\Http\Controllers\API\ArticlePicturesController;
 use App\Http\Controllers\API\PointCategoriesController;
 use App\Http\Controllers\API\ArticleCategoriesController;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -31,57 +32,51 @@ Route::prefix('/categories')->group(function () {
 });
 
 Route::middleware('auth:sanctum')->group(function () {
-    Route::post('/logout', [SecurityController::class, 'logout']);
-    Route::get('/', [UsersController::class, 'index']);
-    Route::post('/users', [UsersController::class, 'store']);
-    Route::get('/{user}', [UsersController::class, 'show']);
-    Route::post('/edit/{user}', [UsersController::class, 'update']);
-    Route::delete('/{user}', [UsersController::class, 'destroy']);
-    Route::post('/articles', [ArticlesController::class, 'store']);
-    Route::post('/edit/{article}', [ArticlesController::class, 'update']);
-    Route::delete('/{article}', [ArticlesController::class, 'destroy']);
-    Route::post('/article-categories', [ArticleCategoriesController::class, 'store']);
-    Route::post('/edit/{article-categorie}', [ArticleCategoriesController::class, 'update']);
-    Route::delete('/{article-categorie}', [ArticleCategoriesController::class, 'destroy']);
-    Route::post('/point-categories', [PointCategoriesController::class, 'store']);
-    Route::post('/edit/{point-categorie}', [PointCategoriesController::class, 'update']);
-    Route::post('/{point-categorie}', [PointCategoriesController::class, 'destroy']);
-    Route::post('/interest-points', [InterestPointsController::class, 'store']);
-    Route::post('/edit/{interest-point}', [InterestPointsController::class, 'update']);
-    Route::delete('/{interest-point}', [InterestPointsController::class, 'destroy']);
-    Route::get('/', [ArticlePicturesController::class, 'index']);
-    Route::post('/article-pictures', [ArticlePicturesController::class, 'store']);
-    Route::get('/{article-picture}', [ArticlePicturesController::class, 'show']);
-    Route::get('/edit/{article-picture}', [ArticlePicturesController::class, 'update']);
-    Route::delete('/{article-picture}', [ArticlePicturesController::class, 'destroy']);
-    Route::get('/', [PointPicturesController::class, 'index']);
-    Route::post('/point-pictures', [PointPicturesController::class, 'store']);
-    Route::get('/{point-picture}', [PointPicturesController::class, 'show']);
-    Route::post('/edit/{point-picture}', [PointPicturesController::class, 'update']);
-    Route::delete('/{point-picture}', [PointPicturesController::class, 'destroy']);
-    Route::get('/', [UsersController::class, 'index']);
-    Route::post('/edit/{user}', [UsersController::class, 'update']);
+    Route::get('/me', function(){
+        return Auth::user();
+    });
 });
 
-// Routes pour les utilisateurs
-Route::prefix('/articles')->group(function () {
+Route::post('/logout', [SecurityController::class, 'logout'])->middleware('auth:sanctum');;
+
+Route::prefix('/users')->group(function(){
+    Route::get('/', [UsersController::class, 'index']);
+    Route::get('/{user}', [UsersController::class, 'show']);
+    Route::post('/', [UsersController::class, 'store'])->middleware('auth:sanctum');
+    Route::post('/edit/{user}', [UsersController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/{user}', [UsersController::class, 'destroy'])->middleware('auth:sanctum');
+});
+
+Route::prefix('/articles')->group(function(){
     Route::get('/', [ArticlesController::class, 'index']);
     Route::get('/{article}', [ArticlesController::class, 'show']);
+    Route::post('/', [ArticlesController::class, 'store'])->middleware('auth:sanctum');
+    Route::post('/edit/{article}', [ArticlesController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/{article}', [ArticlesController::class, 'destroy'])->middleware('auth:sanctum');
 });
 
-Route::prefix('/article-categories')->group(function () {
+Route::prefix('/article-categories')->group(function(){
     Route::get('/', [ArticleCategoriesController::class, 'index']);
-    Route::get('/{categorie}', [ArticleCategoriesController::class, 'show']);
+    Route::get('/{category}', [ArticleCategoriesController::class, 'show']);
+    Route::post('/', [ArticleCategoriesController::class, 'store'])->middleware('auth:sanctum');
+    Route::post('/edit/{category}', [ArticleCategoriesController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/{category}', [ArticleCategoriesController::class, 'destroy'])->middleware('auth:sanctum');
+});
+Route::prefix('/point-categories')->group(function(){
+    Route::get('/', [PointCategoriesController::class, 'index']);
+    Route::get('/{category}', [PointCategoriesController::class, 'show']);
+    Route::post('/', [PointCategoriesController::class, 'store'])->middleware('auth:sanctum');
+    Route::post('/edit/{category}', [PointCategoriesController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/{category}', [PointCategoriesController::class, 'destroy'])->middleware('auth:sanctum');
 });
 
-Route::prefix('/point-categories')->group(function () {
-    Route::get('/', [PointCategoriesController::class, 'index']);
-    Route::get('/{categorie}', [PointCategoriesController::class, 'show']);
-});
 
 Route::prefix('/interest-points')->group(function () {
     Route::get('/', [InterestPointsController::class, 'index']);
     Route::get('/{point}', [InterestPointsController::class, 'show']);
+    Route::post('/interest-points', [InterestPointsController::class, 'store'])->middleware('auth:sanctum');
+    Route::post('/edit/{interest-point}', [InterestPointsController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('/{interest-point}', [InterestPointsController::class, 'destroy'])->middleware('auth:sanctum');
 });
 
 Route::prefix('/security')->group(function () {
